Split out @ember/debug/-basic
