+++ lib/index.js
@@ -204,6 +204,7 @@ module.exports = {
       new Funnel(packages, {
         include: [
           '@ember/debug/lib/**',
+          '@ember/debug/-basic.js',
           '@ember/debug/index.js',
           'ember-testing/index.js',
           'ember-testing/lib/**',
