+++ packages/@ember/debug/lib/inspect.ts
@@ -1,5 +1,5 @@
 import { _WeakSet as WeakSet } from '@glimmer/util';
-import { assert } from '@ember/debug';
+import { assert } from '@ember/debug/-basic';
 const { toString: objectToString } = Object.prototype;
 const { toString: functionToString } = Function.prototype;
 const { isArray } = Array;
