+++ packages/@ember/debug/type-tests/debug.test.ts
@@ -1,4 +1,4 @@
-import { debug } from '@ember/debug';
+import { debug } from '@ember/debug/-basic';
 import { expectTypeOf } from 'expect-type';
 
 expectTypeOf(debug("I'm a debug notice!")).toEqualTypeOf<void>();
